pyproject.toml
[project]
name = "spase-cache"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fitter>=1.8.0",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "omegaconf>=2.3.0",
    "pandas>=3.0.1",
    "polars>=1.39.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
